# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GearUp API",
    "version": "1.0.0",
    "description": "Sports and outdoor gear rental service backend API"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/reviews/gear/{gearId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get reviews for a gear item (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ]
      }
    },
...
  },
...
}